# perf: lazy load treatment for `ape-ethereum` core plugin