### What I did

this plugin is heavily imported from

before

```python
In [1]: %time import ape_ethereum
CPU times: user 2.09 s, sys: 545 ms, total: 2.63 s
Wall time: 1.71 s
```

after

```python
In [1]: %time import ape_ethereum
CPU times: user 8.22 ms, sys: 2.44 ms, total: 10.7 ms
Wall time: 9.53 ms
```

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated